Speedup Metropolis and NUTS

+++ pymc3/distributions/multivariate.py
@@ -222,7 +222,7 @@ class LKJCorr(Continuous):
     def __init__(self, n, p, *args, **kwargs):
         self.n = n
         self.p = p
-        n_elem = p * (p - 1) / 2
+        n_elem = int(p * (p - 1) / 2)
         self.mean = np.zeros(n_elem)
         super(LKJCorr, self).__init__(shape=n_elem, *args, **kwargs)
